Set source of diagnostic to yaml tree element
